[Update] login & register update
