--- frontend/Project/src/data/qnaDummy.js
@@ -1,94 +1,0 @@
-const qnaDummy = [
-  {
-    id: 1,
-    title: "로그인이 안돼요",
-    content: "로그인 버튼을 눌러도 반응이 없습니다. 어떻게 해야 하나요?",
-    author: "user123",
-    date: "2025-08-01",
-    answer: "캐시를 삭제한 뒤 다시 로그인 시도해보세요.",
-    answeredAt: "2025-08-03",
-    answeredBy: "admin",
-  },
-  {
-    id: 2,
-    title: "비밀번호는 어떻게 바꾸나요?",
-    content: "비밀번호 변경 메뉴를 찾을 수 없습니다.",
-    author: "user456",
-    date: "2025-08-02",
-    answer: "",
-    answeredAt: null,
-    answeredBy: null,
-  },
-  {
-    id: 3,
-    title: "회원탈퇴는 어디서 하나요?",
-    content: "회원 탈퇴는 어디에서 진행할 수 있나요?",
-    author: "user789",
-    date: "2025-08-05",
-    answer: "마이페이지 하단에서 회원탈퇴를 진행하실 수 있습니다.",
-    answeredAt: "2025-08-06",
-    answeredBy: "admin",
-  },
-  {
-    id: 4,
-    title: "배송은 보통 얼마나 걸리나요?",
-    content: "서울 기준으로 주문 후 며칠 내에 도착하나요?",
-    author: "user001",
-    date: "2025-08-03",
-    answer: "보통 1~2일 내에 배송됩니다.",
-    answeredAt: "2025-08-04",
-    answeredBy: "staff01",
-  },
-  {
-    id: 5,
-    title: "향수 샘플 신청은 어디서 해요?",
-    content: "무료 샘플을 신청하고 싶은데, 어디서 가능한가요?",
-    author: "user222",
-    date: "2025-08-01",
-    answer: "",
-    answeredAt: null,
-    answeredBy: null,
-  },
-  {
-    id: 6,
-    title: "할인 쿠폰은 어떻게 받나요?",
-    content: "신규 회원 할인 쿠폰이 있다고 들었습니다.",
-    author: "newbie",
-    date: "2025-07-30",
-    answer: "회원 가입 시 자동 발급되며, 마이페이지에서 확인 가능합니다.",
-    answeredAt: "2025-08-01",
-    answeredBy: "admin",
-  },
-  {
-    id: 7,
-    title: "결제 수단에는 뭐가 있나요?",
-    content: "카카오페이나 네이버페이도 가능한가요?",
-    author: "shopper1",
-    date: "2025-07-29",
-    answer: "네, 다양한 간편결제를 지원하고 있습니다.",
-    answeredAt: "2025-07-30",
-    answeredBy: "staff02",
-  },
-  {
-    id: 8,
-    title: "제품 교환은 어떻게 하나요?",
-    content: "불량품이 와서 교환하고 싶습니다.",
-    author: "user654",
-    date: "2025-07-28",
-    answer: "",
-    answeredAt: null,
-    answeredBy: null,
-  },
-  {
-    id: 9,
-    title: "문의한 내용은 어디서 확인하나요?",
-    content: "작성한 Q&A를 다시 확인할 수 있나요?",
-    author: "user999",
-    date: "2025-07-27",
-    answer: "마이페이지 > 나의 문의 내역에서 확인 가능합니다.",
-    answeredAt: "2025-07-28",
-    answeredBy: "admin",
-  },
-];
-
-export default qnaDummy;
