--- frontend/Project/src/page/LoginPage.jsx
@@ -1,14 +1,0 @@
-import LoginComponent from "../component/login/LoginComponent";
-import Footer from "../include/Footer";
-import Header from "../include/Header";
-
-const LoginPage = () => {
-  return (
-    <>
-      <Header />
-      <LoginComponent />
-      <Footer />
-    </>
-  );
-};
-export default LoginPage;
