--- frontend/Project/src/page/RegisterPage.jsx
@@ -1,14 +1,0 @@
-import RegisterComponent from "../component/login/RegisterComponent";
-import Footer from "../include/Footer";
-import Header from "../include/Header";
-
-const RegisterPage = () => {
-  return (
-    <>
-      <Header />
-      <RegisterComponent />
-      <Footer />
-    </>
-  );
-};
-export default RegisterPage;
