+++ frontend/Project/src/page/login/LoginPage.jsx
@@ -1,0 +1,14 @@
+import LoginComponent from "../../component/login/LoginComponent";
+import Footer from "../../include/Footer";
+import Header from "../../include/Header";
+
+const LoginPage = () => {
+  return (
+    <>
+      <Header />
+      <LoginComponent />
+      <Footer />
+    </>
+  );
+};
+export default LoginPage;
