+++ frontend/Project/src/page/login/RegisterPage.jsx
@@ -1,0 +1,14 @@
+import RegisterComponent from "../../component/login/RegisterComponent";
+import Footer from "../../include/Footer";
+import Header from "../../include/Header";
+
+const RegisterPage = () => {
+  return (
+    <>
+      <Header />
+      <RegisterComponent />
+      <Footer />
+    </>
+  );
+};
+export default RegisterPage;
